@@ -1,29 +1,25 @@
 buildscript {
     repositories {
-        jcenter()
         google()
         mavenCentral()
     }
 
     dependencies {
         classpath "com.android.tools.build:gradle:$gradleBuildTools"
-        classpath "com.monits:static-code-analysis-plugin:$staticAnalysis"
     }
 }
-
 
 apply from: rootProject.file('tools.gradle')
 
 allprojects {
     repositories {
-        jcenter()
         google()
         mavenCentral()
     }
 
     tasks.withType(JavaCompile) {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
     }
 
     group = libraryGroupId
