@@ -1,9 +1,7 @@
 apply plugin: 'com.android.application'
 
-apply from: rootProject.file('jacoco.gradle')
-
 android {
-    compileSdkVersion 29
+    compileSdkVersion maxApiLevel.toInteger()
 
     lintOptions{
         // TODO: Delete this once play services is updated. Now we need it because play services brings support library 26
@@ -12,8 +10,8 @@ android {
 
     defaultConfig {
         applicationId "com.mercadolibre.android.device.testapp"
-        minSdkVersion 19
-        targetSdkVersion 29
+        minSdkVersion minApiLevel
+        targetSdkVersion maxApiLevel
         versionCode 1
         versionName "1.0"
 
@@ -29,13 +27,14 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility = 1.8
-        targetCompatibility = 1.8
+        sourceCompatibility = 11
+        targetCompatibility = 11
     }
 }
 
 dependencies {
     implementation project(":sdk")
-    implementation "com.android.support:appcompat-v7:$support"
-    implementation "com.android.support:design:$support"
+
+    implementation "androidx.appcompat:appcompat:$androidXAppCompat"
+    implementation "com.google.android.material:material:$googleMaterial"
 }
